draft JavaScript file of Facebook Api
This is a draft JavaScript file of Facebook Api. 
(pseudocode) It will connect to the Facebook sdk and api, and runs through the comments. If they are negative are negative comments or have profanity harassing the user, they will move to the possibleBlock which are divided into two levels according to the number of the negative comments.